# lesson-12-exam/src/api/v1/conditions/condition.model.ts
export interface ConditionEntity extends CreateCondition {
    id: number;
}

export interface CreateCondition {
    month: number;
    percent: number;
    agreement: string;
}

export interface UpdateCondition extends Partial<CreateCondition> { }