# lesson-12-exam/src/api/v1/products/product.model.ts
export interface ProductEntity extends CreateProduct {
    id: number;
}

export interface CreateProduct {
    name: string;
    price: number;
    count: number;
};

export interface UpdateProduct extends Partial<CreateProduct> { };