# lesson-12-exam/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"StatusEntity": {
				"properties": {
					"contract": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"is_completed": {
						"type": "boolean"
					}
				},
				"required": [
					"contract",
					"id",
					"is_completed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductEntity": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"price",
					"count",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProduct": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"price",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProduct": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PaymentEntity": {
				"properties": {
					"amount": {
						"type": "number",
						"format": "double"
					},
					"contract": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"is_paid": {
						"type": "boolean"
					}
				},
				"required": [
					"amount",
					"contract",
					"id",
					"is_paid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePayment": {
				"properties": {
					"amount": {
						"type": "number",
						"format": "double"
					},
					"contract": {
						"type": "number",
						"format": "double"
					},
					"is_paid": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CustomerEntity": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"passport_series": {
						"type": "string"
					},
					"passport_number": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"phone",
					"passport_series",
					"passport_number",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCustomer": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"passport_series": {
						"type": "string"
					},
					"passport_number": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"phone",
					"passport_series",
					"passport_number"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCustomer": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"passport_series": {
						"type": "string"
					},
					"passport_number": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ContractEntity": {
				"properties": {
					"customer": {
						"type": "number",
						"format": "double"
					},
					"product": {
						"type": "number",
						"format": "double"
					},
					"agreement": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"customer",
					"product",
					"agreement",
					"id",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateContract": {
				"properties": {
					"customer": {
						"type": "number",
						"format": "double"
					},
					"product": {
						"type": "number",
						"format": "double"
					},
					"agreement": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"customer",
					"product",
					"agreement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConditionEntity": {
				"properties": {
					"month": {
						"type": "number",
						"format": "double"
					},
					"percent": {
						"type": "number",
						"format": "double"
					},
					"agreement": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"month",
					"percent",
					"agreement",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCondition": {
				"properties": {
					"month": {
						"type": "number",
						"format": "double"
					},
					"percent": {
						"type": "number",
						"format": "double"
					},
					"agreement": {
						"type": "string"
					}
				},
				"required": [
					"month",
					"percent",
					"agreement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCondition": {
				"properties": {
					"month": {
						"type": "number",
						"format": "double"
					},
					"percent": {
						"type": "number",
						"format": "double"
					},
					"agreement": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "exam",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/v1/status/all": {
			"get": {
				"operationId": "GET_ALL",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/StatusEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Statuses"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/status/{id}": {
			"get": {
				"operationId": "GET_BY_ID",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StatusEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Statuses"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/product": {
			"post": {
				"operationId": "CREATE",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProduct"
							}
						}
					}
				}
			}
		},
		"/v1/product/all": {
			"get": {
				"operationId": "GET_ALL",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/product/{id}": {
			"put": {
				"operationId": "UPDATE",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProduct"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GET_BY_ID",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DELETE",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"type": "boolean"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/product/sold": {
			"get": {
				"operationId": "GetSoldProducts",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"default": "2025-05-27T16:46:12.312Z",
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/product/overdue": {
			"get": {
				"operationId": "GET_OVERDUE",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/payment/{id}": {
			"put": {
				"operationId": "UPDATE",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePayment"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GET_BY_ID",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/payment/all": {
			"get": {
				"operationId": "GET_ALL",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PaymentEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/customer": {
			"post": {
				"operationId": "CREATE",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCustomer"
							}
						}
					}
				}
			}
		},
		"/v1/customer/all": {
			"get": {
				"operationId": "GET_ALL",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CustomerEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/customer/{id}": {
			"put": {
				"operationId": "UPDATE",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCustomer"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DELETE",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"type": "boolean"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"get": {
				"operationId": "GET_BY_ID",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Customers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/contract": {
			"post": {
				"operationId": "CREATE",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContractEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Contracts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateContract"
							}
						}
					}
				}
			}
		},
		"/contract/all": {
			"get": {
				"operationId": "GET_ALL",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ContractEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Contracts"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/contract/{id}": {
			"get": {
				"operationId": "GET_BY_ID",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContractEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Contracts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/condition": {
			"post": {
				"operationId": "CREATE",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConditionEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCondition"
							}
						}
					}
				}
			}
		},
		"/v1/condition/all": {
			"get": {
				"operationId": "GET_ALL",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ConditionEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/condition/{id}": {
			"get": {
				"operationId": "GET_BY_ID",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConditionEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UPDATE",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConditionEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCondition"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DELETE",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"type": "boolean"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}